elasticsearch/requirements.txt
flask==2.1.0
elasticsearch==7.15.0
openai==0.10.2
transformers==4.12.0
python-dotenv==0.19.1
